Use org.kie.kogito package prefix - dmn-pmml-quarkus
