+++ dmn-pmml-quarkus-example/src/test/java/org/kie/kogito/dmn/pmml/quarkus/example/CommonTestUtils.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kie.dmn.pmml.kogito.quarkus.example;
+package org.kie.kogito.dmn.pmml.quarkus.example;
 
 import java.util.Map;
