+++ dmn-pmml-quarkus-example/src/test/java/org/kie/kogito/dmn/pmml/quarkus/example/DMNRegressionTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 Red Hat, Inc. and/or its affiliates.
+ * Copyright 2020 Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kie.dmn.pmml.kogito.quarkus.example;
+package org.kie.kogito.dmn.pmml.quarkus.example;
 
 import org.junit.jupiter.api.Test;
 
@@ -24,31 +24,33 @@ import io.restassured.path.json.config.JsonPathConfig;
 
 import static io.restassured.RestAssured.given;
 import static io.restassured.config.JsonConfig.jsonConfig;
-import static org.hamcrest.Matchers.*;
+import static org.hamcrest.Matchers.closeTo;
+import static org.hamcrest.Matchers.comparesEqualTo;
+import static org.hamcrest.Matchers.is;
 
 @QuarkusTest
-public class DMNMiningModelTest {
+public class DMNRegressionTest {
 
     static {
         RestAssured.enableLoggingOfRequestAndResponseIfValidationFails();
     }
 
     @Test
-    public void testEvaluateMiningModelDMN() {
-        String inputData = "{\"input1\":200.0, \"input2\":-1.0, \"input3\":2.0}";
+    public void testEvaluateRegressionDMN() {
+        String inputData = "{\"fld1\":3.0, \"fld2\":2.0, \"fld3\":\"y\"}";
         given()
                 .config(RestAssured.config().jsonConfig(jsonConfig().numberReturnType(JsonPathConfig.NumberReturnType.DOUBLE)))
                 .contentType(ContentType.JSON)
                 .body(inputData)
                 .when()
-                .post("/TestMiningModelDMN")
+                .post("/TestRegressionDMN")
                 .then()
                 .statusCode(200)
-                .body("SumMiningModelBKM", is("function SumMiningModelBKM( input1, input2, input3 )"))
-                .body("input1", is(comparesEqualTo(200))) // was input
-                .body("input2", is(comparesEqualTo(-1))) // was input
-                .body("input3", is(comparesEqualTo(2))) // was input
-                .body("Decision", is(comparesEqualTo(-299))) // real decision output
+                .body("RegressionModelBKM", is("function RegressionModelBKM( fld1, fld2, fld3 )"))
+                .body("fld3", is("y")) // was input
+                .body("fld2", is(comparesEqualTo(2))) // was input
+                .body("fld1", is(comparesEqualTo(3))) // was input
+                .body("Decision", is(closeTo(52.5, 0))) // real decision output 
         ;
     }
 }
