+++ dmn-pmml-quarkus-example/src/test/java/org/kie/kogito/dmn/pmml/quarkus/example/DMNTreeTest.java
@@ -13,44 +13,37 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kie.dmn.pmml.kogito.quarkus.example;
+package org.kie.kogito.dmn.pmml.quarkus.example;
 
 import org.junit.jupiter.api.Test;
 
 import io.quarkus.test.junit.QuarkusTest;
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
-import io.restassured.path.json.config.JsonPathConfig;
 
 import static io.restassured.RestAssured.given;
-import static io.restassured.config.JsonConfig.jsonConfig;
-import static org.hamcrest.Matchers.closeTo;
-import static org.hamcrest.Matchers.comparesEqualTo;
 import static org.hamcrest.Matchers.is;
 
 @QuarkusTest
-public class DMNRegressionTest {
+public class DMNTreeTest {
 
     static {
         RestAssured.enableLoggingOfRequestAndResponseIfValidationFails();
     }
 
     @Test
-    public void testEvaluateRegressionDMN() {
-        String inputData = "{\"fld1\":3.0, \"fld2\":2.0, \"fld3\":\"y\"}";
+    public void testEvaluateTreeDMN() {
+        String inputData = "{\"temperature\":30.0, \"humidity\":10.0}";
         given()
-                .config(RestAssured.config().jsonConfig(jsonConfig().numberReturnType(JsonPathConfig.NumberReturnType.DOUBLE)))
                 .contentType(ContentType.JSON)
                 .body(inputData)
                 .when()
-                .post("/TestRegressionDMN")
+                .post("/TestTreeDMN")
                 .then()
                 .statusCode(200)
-                .body("RegressionModelBKM", is("function RegressionModelBKM( fld1, fld2, fld3 )"))
-                .body("fld3", is("y")) // was input
-                .body("fld2", is(comparesEqualTo(2))) // was input
-                .body("fld1", is(comparesEqualTo(3))) // was input
-                .body("Decision", is(closeTo(52.5, 0))) // real decision output 
-        ;
+                .body("TestTreeBKM", is("function TestTreeBKM( humidity, temperature )"))
+                .body("temperature", is(Float.valueOf("30")))
+                .body("humidity", is(Float.valueOf("10")))
+                .body("Decision", is("sunglasses"));
     }
 }
