+++ dmn-pmml-quarkus-example/src/test/java/org/kie/kogito/dmn/pmml/quarkus/example/ScoreCardTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kie.dmn.pmml.kogito.quarkus.example;
+package org.kie.kogito.dmn.pmml.quarkus.example;
 
 import java.util.Collections;
 import java.util.Map;
@@ -23,8 +23,8 @@ import org.junit.jupiter.api.Test;
 import io.quarkus.test.junit.QuarkusTest;
 import io.restassured.RestAssured;
 
-import static org.kie.dmn.pmml.kogito.quarkus.example.CommonTestUtils.testDescriptive;
-import static org.kie.dmn.pmml.kogito.quarkus.example.CommonTestUtils.testResult;
+import static org.kie.kogito.dmn.pmml.quarkus.example.CommonTestUtils.testDescriptive;
+import static org.kie.kogito.dmn.pmml.quarkus.example.CommonTestUtils.testResult;
 
 @QuarkusTest
 public class ScoreCardTest {
